requirements.txt
streamlit
nltk
textstat
python-docx
striprtf
reportlab
